third_party/WebKit/LayoutTests/bluetooth/requestDevice/acceptAllDevices/device-with-empty-name.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script src="../../../resources/mojo-helpers.js"></script>
<script src="../../../resources/bluetooth/web-bluetooth-test.js"></script>
<script>
'use strict';
promise_test(() => {
  return setUpPreconnectedDevice({name: ''})
    .then(() => requestDeviceWithKeyDown({acceptAllDevices: true}))
    .then(device => {
      assert_equals(device.name, '');
    });
}, 'Device with empty name and no UUIDs nearby. Should be found if ' +
   'acceptAllDevices is true.');
</script>
